Finished renaming from neuroport_dbs to open_mer. Unfortunately, git didn't recognize it as a move so it's a huge delete and addition.
