+++ open_mer/dbsgui/widgets/ui/cbsdk_connect.ui
