+++ open_mer/dbsgui/widgets/ui/send_comments.ui
