+++ open_mer/resources/config/WaveformGUI.ini
