+++ open_mer/resources/icons/depth_status_in_use.png
