+++ open_mer/resources/icons/depth_status_off.png
