+++ open_mer/scripts/DepthGUI.py
@@ -1,14 +1,15 @@
 import sys
 from qtpy import QtCore, QtWidgets
-from neuroport_dbs.dbsgui.waveform import WaveformGUI
+from open_mer.dbsgui.depth import DepthGUI
 
 
 def main():
     _ = QtWidgets.QApplication(sys.argv)
-    aw = WaveformGUI()
+    window = DepthGUI()
+    window.show()
     timer = QtCore.QTimer()
-    timer.timeout.connect(aw.update)
-    timer.start(1)
+    timer.timeout.connect(window.update)
+    timer.start(100)
 
     if (sys.flags.interactive != 1) or not hasattr(QtCore, 'PYQT_VERSION'):
         QtWidgets.QApplication.instance().exec_()
