+++ open_mer/scripts/MappingGUI.py
@@ -1,15 +1,15 @@
 import sys
 import qtpy
-from neuroport_dbs.dbsgui.sweep import SweepGUI
+from qtpy import QtWidgets, QtCore
+from open_mer.dbsgui.mapping import MappingGUI
 
 
 def main():
-    from qtpy import QtWidgets, QtCore
     _ = QtWidgets.QApplication(sys.argv)
-    aw = SweepGUI()
-    timer = QtCore.QTimer()
-    timer.timeout.connect(aw.update)
-    timer.start(1)
+    aw = MappingGUI()
+    # timer = QTimer()
+    # timer.timeout.connect(aw.update)
+    # timer.start(1000)
 
     if (sys.flags.interactive != 1) or not hasattr(qtpy.QtCore, 'PYQT_VERSION'):
         QtWidgets.QApplication.instance().exec_()
