+++ setup.cfg
@@ -1,14 +1,14 @@
 [metadata]
-name = neuroport_dbs
-version = attr: neuroport_dbs.__version__
+name = open_mer
+version = attr: open_mer.__version__
 author = Chadwick Boulay
 author_email = chboulay@ohri.ca
-description = Tools for clinical research using Blackrock Neuroport in DBS MER
+description = Tools for deep brain stimulation (DBS) surgery intraoperative mapping with microelectrode recording (MER)
 long_description = file: docs/README.md
 long_description_content_type = text/markdown
-url = https://github.com/SachsLab/NeuroportDBS
+url = https://github.com/SachsLab/OpenMER
 project_urls =
-    Bug Tracker = https://github.com/SachsLab/NeuroportDBS/issues
+    Bug Tracker = https://github.com/SachsLab/OpenMER/issues
 classifiers =
     Programming Language :: Python :: 3
     License :: OSI Approved :: Apache Software License
@@ -43,15 +43,15 @@ exclude =
     docs
 
 [options.package_data]
-"neuroport_dbs/resources/icons" = "*.png"
-"neuroport_dbs/resources/config" = "*.ini"
+"open_mer/resources/icons" = "*.png"
+"open_mer/resources/config" = "*.ini"
 
 [options.entry_points]  # https://setuptools.pypa.io/en/latest/userguide/entry_point.html
 gui_scripts =
-    dbs-sweep = neuroport_dbs.scripts.SweepGUI:main
-    dbs-waveform = neuroport_dbs.scripts.WaveformGUI:main
-    dbs-raster = neuroport_dbs.scripts.RasterGUI:main
-    dbs-features = neuroport_dbs.scripts.FeaturesGUI:main
-    dbs-mapping = neuroport_dbs.scripts.MappingGUI:main
-    dbs-comments = neuroport_dbs.scripts.CommentsGUI:main
-    dbs-ddu = neuroport_dbs.scripts.DepthGUI:main
+    dbs-sweep = open_mer.scripts.SweepGUI:main
+    dbs-waveform = open_mer.scripts.WaveformGUI:main
+    dbs-raster = open_mer.scripts.RasterGUI:main
+    dbs-features = open_mer.scripts.FeaturesGUI:main
+    dbs-mapping = open_mer.scripts.MappingGUI:main
+    dbs-comments = open_mer.scripts.CommentsGUI:main
+    dbs-ddu = open_mer.scripts.DepthGUI:main
